LPD-32807 Move StyleBookUtil to style-book-api so it can be shared externally.
